requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pydantic-settings
passlib[bcrypt]
PyJWT
bcrypt<4
python-dotenv
SQLAlchemy
email-validator
alembic
